separate FlowerServiceRunnable from MainActivity.kt
